Create separate release and publish processes.
