+++ tasks/release.js
@@ -1,12 +1,8 @@
 "use strict"
 
 const fs = require("fs")
-const path = require("path")
-const toml = require("toml")
-const { execSync } = require("child_process")
 const git = require("simple-git/promise")()
-const release = require("publish-release")
-const util = require("util")
+const octokit = require(`@octokit/rest`)()
 
 function bumpVersion(versionString) {
   let versionElements = versionString.split(".")
@@ -26,7 +22,7 @@ function updateChangeLog(changeLog, newVersion, now) {
 const updatePackageJson = (packageJson, version) =>
   Object.assign({}, packageJson, { version })
 
-const pushCommit = async (token, tag) => {
+const pushCommit = async (token, tag, head) => {
   await Promise.all([
     git.addConfig("user.name", "Voyager Bot"),
     git.addConfig("user.email", "voyager_bot@tendermint.com")
@@ -42,57 +38,31 @@ const pushCommit = async (token, tag) => {
   // needed to authenticate properly
   await git.addRemote("bot", `https://${token}@github.com/cosmos/voyager.git`)
 
-  await git.push("bot", "HEAD:master")
+  await git.push("bot", `HEAD:${head}`, { "--tags": null })
 }
 
-const build = defaultNetwork => {
-  console.log("--- BUILDING ---")
-
-  execSync(
-    `node tasks/build/build.js \
-      --network=${path.join(__dirname, `../app/networks`, defaultNetwork)}`,
-    {
-      stdio: `inherit`
-    }
-  )
-
-  console.log("--- DONE BUILDING ---")
-}
-
-const assetsDir = path.join(__dirname, `../builds/Voyager`)
-
-const publishRelease = (token, tag) =>
-  util.promisify(release)({
-    token,
-    owner: "cosmos",
-    repo: "voyager",
-    tag,
-    name: `Cosmos Voyager Alpha ${tag} (UNSAFE)`,
-    notes: `
-NOTE: DO NOT ENTER YOUR FUNDRAISER SEED. THIS SOFTWARE HAS NOT BEEN AUDITED.
-NEVER ENTER YOUR FUNDRAISER SEED 12 WORDS ONTO AN ONLINE COMPUTER.
-
-Even when we do start supporting fundraiser seeds, don't use it except for
-testing or with small amounts. We will release a CLI to use for offline signing
-of transactions, and we will also add hardware support for this UI.
-
-Please checkout the [CHANGELOG.md](CHANGELOG.md) for a list of changes.
-`,
-    prerelease: true,
-    assets: fs.readdirSync(assetsDir).map(file => path.join(assetsDir, file))
+const createPullRequest = async (token, tag, head) => {
+  octokit.authenticate({
+    type: `token`,
+    token
   })
 
+  await octokit.pullRequests.create({
+    owner: `cosmos`,
+    repo: `voyager`,
+    title: `automatic release created for ${tag}`,
+    head,
+    base: `develop`,
+    maintainer_can_modify: true
+  })
+}
+
 async function main() {
-  console.log("Releasing Voyager...")
+  console.log("Making release...")
   const changeLog = fs.readFileSync(__dirname + "/../CHANGELOG.md", "utf8")
   const packageJson = require(__dirname + "/../package.json")
   const oldVersion = packageJson.version
   const newVersion = bumpVersion(oldVersion)
-
-  const config = toml.parse(
-    fs.readFileSync(__dirname + "/../app/config.toml", "utf8")
-  )
-
   console.log("New version:", newVersion)
   const newChangeLog = updateChangeLog(changeLog, newVersion, new Date())
   const newPackageJson = updatePackageJson(packageJson, newVersion)
@@ -107,11 +77,9 @@ async function main() {
 
   console.log("--- Committing release changes ---")
   const tag = `v${newVersion}`
-  await pushCommit(process.env.GIT_BOT_TOKEN, tag)
-  build(config.default_network)
-  console.log("--- Publishing release ---")
-  await publishRelease(process.env.GIT_BOT_TOKEN, tag)
-  console.log("--- Done releasing ---")
+  const head = `release/${tag}`
+  await pushCommit(process.env.GIT_BOT_TOKEN, tag, head)
+  await createPullRequest(process.env.GIT_BOT_TOKEN, tag, head)
 }
 
 if (require.main === module) {
